Increase max content size

+++ src/api.rs
@@ -228,7 +228,7 @@ pub async fn stream_response<'a>(
             };
 
             CreateChatCompletionRequestArgs::default()
-                .max_tokens(config.context_max_tokens)
+                .max_tokens(config.context_max_tokens as u32)
                 .model(model)
                 .stop(config.stop_words.clone())
                 .messages(full_conversation)
@@ -236,7 +236,7 @@ pub async fn stream_response<'a>(
                 .build()?
         }
         None => CreateChatCompletionRequestArgs::default()
-            .max_tokens(config.context_max_tokens)
+            .max_tokens(config.context_max_tokens as u32)
             .model(model)
             .stop(config.stop_words.clone())
             .messages(full_conversation)
@@ -374,7 +374,7 @@ pub async fn fetch_response<'a>(
             };
 
             CreateChatCompletionRequestArgs::default()
-                .max_tokens(config.context_max_tokens)
+                .max_tokens(config.context_max_tokens as u32)
                 .model(model)
                 .stop(config.stop_words.clone())
                 .messages(full_conversation)
@@ -382,7 +382,7 @@ pub async fn fetch_response<'a>(
                 .build()?
         }
         None => CreateChatCompletionRequestArgs::default()
-            .max_tokens(config.context_max_tokens)
+            .max_tokens(config.context_max_tokens as u32)
             .model(model)
             .stop(config.stop_words.clone())
             .messages(full_conversation)
